requirements.txt
python-telegram-bot
yt-dlp
aiohttp
python-dotenv
